+++ changelog.md
@@ -1,3 +1,4 @@
 # v1.1.0
 
+ * Add support for importing/exporting lists as `.gmdl` files
  * Allow importing multiple levels at once